% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{as_g6_node}
\alias{as_g6_node}
\alias{as_g6_node.g6_node}
\alias{as_g6_node.list}
\alias{as_g6_edge}
\alias{as_g6_edge.g6_edge}
\alias{as_g6_edge.list}
\alias{as_g6_combo}
\alias{as_g6_combo.g6_combo}
\alias{as_g6_combo.list}
\title{Coerce to a g6 element object}
\usage{
as_g6_node(x, ...)

\method{as_g6_node}{g6_node}(x, ...)

\method{as_g6_node}{list}(x, ...)

as_g6_edge(x, ...)

\method{as_g6_edge}{g6_edge}(x, ...)

\method{as_g6_edge}{list}(x, ...)

as_g6_combo(x, ...)

\method{as_g6_combo}{g6_combo}(x, ...)

\method{as_g6_combo}{list}(x, ...)
}
\arguments{
\item{x}{An object to be coerced. List or g6 element are supported.}

\item{...}{Additional arguments (unused).}
}
\value{
An element of class \code{g6_node}, \code{g6_edge}, or \code{g6_combo}.
}
\description{
Coerce to a g6 element object
}
