% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{as_g6_nodes}
\alias{as_g6_nodes}
\alias{as_g6_edges}
\alias{as_g6_edges.g6_edges}
\alias{as_g6_edges.data.frame}
\alias{as_g6_edges.list}
\alias{as_g6_combos}
\alias{as_g6_combos.g6_combos}
\alias{as_g6_combos.data.frame}
\alias{as_g6_combos.list}
\title{Coerce to a list of g6_elements objects}
\usage{
as_g6_nodes(x, ...)

as_g6_edges(x, ...)

\method{as_g6_edges}{g6_edges}(x, ...)

\method{as_g6_edges}{data.frame}(x, ...)

\method{as_g6_edges}{list}(x, ...)

as_g6_combos(x, ...)

\method{as_g6_combos}{g6_combos}(x, ...)

\method{as_g6_combos}{data.frame}(x, ...)

\method{as_g6_combos}{list}(x, ...)
}
\arguments{
\item{x}{An object to be coerced. Data frame, list
or g6 element are supported.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{g6_nodes}, \code{g6_edges}, or \code{g6_combos}.
}
\description{
Coerce to a list of g6_elements objects
}
