% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{as_g6_port}
\alias{as_g6_port}
\alias{as_g6_port.g6_port}
\alias{as_g6_port.list}
\title{Coerce to a g6_port object}
\usage{
as_g6_port(x, ...)

\method{as_g6_port}{g6_port}(x, ...)

\method{as_g6_port}{list}(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{g6_port}.
}
\description{
Coerce to a g6_port object
}
\examples{
as_g6_port(list(key = "input-1", type = "input", placement = "left"))
as_g6_port(g6_port("input-1", type = "input"))
}
