% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{as_g6_ports}
\alias{as_g6_ports}
\alias{as_g6_ports.g6_ports}
\alias{as_g6_ports.list}
\title{Coerce to a list of g6_port objects (g6_ports)}
\usage{
as_g6_ports(x, ...)

\method{as_g6_ports}{g6_ports}(x, ...)

\method{as_g6_ports}{list}(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{g6_ports}.
}
\description{
Coerce to a list of g6_port objects (g6_ports)
}
\examples{
as_g6_ports(list(
  list(key = "input-1", type = "input", placement = "left"),
  list(key = "output-1", type = "output", placement = "right")
))
as_g6_ports(g6_ports(
  g6_port("input-1", type = "input"),
  g6_port("output-1", type = "output")
))
}
