% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dag}
\alias{dag}
\title{Example DAG graph}
\format{
A list with 2 data frames:
\describe{
\item{nodes}{data frame with 12 rows for the nodes.}
\item{edges}{data frame with 12 rows for the edges.}
\item{combo}{data frame with 3 rows for the combos.}
}
}
\source{
\url{https://gw.alipayobjects.com/os/antvdemo/assets/data/algorithm-category.json}
}
\usage{
data(dag)
}
\description{
The graph is a directed acyclic graph
}
\keyword{datasets}
