% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{g6_state}
\alias{g6_state}
\alias{g6_nodes}
\alias{g6_nodes_ids}
\alias{g6_edges}
\alias{g6_edges_ids}
\alias{g6_combos}
\alias{g6_combos_ids}
\title{Get state of g6 instance}
\usage{
g6_state(graph)

g6_nodes(graph)

g6_nodes_ids(graph)

g6_edges(graph)

g6_edges_ids(graph)

g6_combos(graph)

g6_combos_ids(graph)
}
\arguments{
\item{graph}{A g6_proxy object created with \code{\link{g6_proxy}}.}
}
\value{
A list containing the state of the g6 instance,
which includes nodes, edges, and combos.

A list of nodes in the g6 instance.

A vector of node ids in the g6 instance.

A list of edges in the g6 instance.

A vector of edge ids in the g6 instance.

A list of combos in the g6 instance.

A vector of combo ids in the g6 instance.
}
\description{
These functions query the current state of a g6 instance in a Shiny app to
extract information about nodes, edges, and combos.
}
\note{
Only works from the server of a Shiny app.
}
