% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{g6_add_edges}
\alias{g6_add_edges}
\title{Add edges to a g6 graph via proxy}
\usage{
g6_add_edges(graph, edges)
}
\arguments{
\item{graph}{A g6_proxy object created with \code{\link{g6_proxy}}.}

\item{edges}{A data frame or list specifying the edges to be added.
If a data frame is provided, each row will be converted to an edge.
Each edge should have at least an 'id' field, and may include other attributes
such as 'label', 'style', 'x', 'y', etc. based on G6 edge specifications.}
}
\value{
The g6_proxy object (invisibly), allowing for method chaining.
}
\description{
This function adds one or more edges to an existing g6 graph instance
using a proxy object. This allows updating the graph without completely
re-rendering it.
}
\details{
This function can only be used with a g6_proxy object within a Shiny application.
It will not work with regular g6 objects outside of Shiny.

If a node with the same ID already exists, it will not be added again.
See \url{https://g6.antv.antgroup.com/en/api/data#graphaddnodedata} for more details.
}
\seealso{
\code{\link{g6_proxy}}, \code{\link{g6_remove_edges}}
}
