% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{g6_data}
\alias{g6_data}
\title{Create a g6_data object}
\usage{
g6_data(graph, el, action, type)
}
\arguments{
\item{nodes}{Nodes data.}

\item{edges}{Edges data.}

\item{combos}{Combo data.}
}
\description{
Create compatible data structure for \code{\link[=g6_add_data]{g6_add_data()}}, \code{\link[=g6_set_data]{g6_set_data()}} or
simply \code{\link[=g6]{g6()}}.
}
