% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{g6_ports}
\alias{g6_ports}
\title{Create a List of G6 Ports}
\usage{
g6_ports(...)
}
\arguments{
\item{...}{One or more g6_port objects.}
}
\value{
An S3 object of class 'g6_ports'.
}
\description{
Create a List of G6 Ports
}
\examples{
g6_ports(
  g6_port("input-1", type = "input", placement = "left"),
  g6_port("output-1", type = "output", placement = "right")
)
}
