% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{g6_proxy}
\alias{g6_proxy}
\title{Create a proxy object to modify an existing g6 graph instance}
\usage{
g6_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Character string matching the ID of the g6 graph instance to be modified.}

\item{session}{The Shiny session object within which the graph exists.
By default, this uses the current reactive domain.}
}
\value{
A proxy object of class "g6_proxy" that can be used with g6 proxy methods
such as \code{g6_add_nodes()}, \code{g6_remove_nodes()}, etc.
}
\description{
This function creates a proxy object that can be used to update an existing g6 graph
instance after it has been rendered in the UI. The proxy allows for server-side
modifications of the graph without completely re-rendering it.
}
