% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{g6_remove_combos}
\alias{g6_remove_combos}
\title{Remove combos from a g6 graph via proxy}
\usage{
g6_remove_combos(graph, ids)
}
\arguments{
\item{graph}{A g6_proxy object created with \code{\link{g6_proxy}}.}

\item{ids}{Character vector or list containing the IDs of the combos to be removed.
If a single ID is provided, it will be converted to a list internally.}
}
\value{
The g6_proxy object (invisibly), allowing for method chaining.
}
\description{
This function removes one or more combos from an existing g6 graph instance
using a proxy object. This allows updating the graph without completely
re-rendering it. When a combo is removed, its connected combos are also removed.
}
\details{
This function can only be used with a g6_proxy object within a Shiny application.
It will not work with regular g6 objects outside of Shiny.

See \url{https://g6.antv.antgroup.com/en/api/data#graphremovecombodata} for more details.
}
\seealso{
\code{\link{g6_proxy}}
}
