% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{g6_update_layout}
\alias{g6_update_layout}
\title{Execute layout for a g6 graph via proxy}
\usage{
g6_update_layout(graph, ...)
}
\arguments{
\item{graph}{A g6_proxy object created with \code{\link{g6_proxy}}.}

\item{...}{Any option to pass to the layout. If so, the layout is updated
before running it.}
}
\value{
The g6_proxy object (invisibly), allowing for method chaining.
}
\description{
This function order the execution of the layout of the current
graph. It can also update layout options before running it.
}
\details{
This function can only be used with a g6_proxy object within a Shiny application.
It will not work with regular g6 objects outside of Shiny.
}
\seealso{
\code{\link{g6_proxy}}
}
