% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{g6_update_nodes}
\alias{g6_update_nodes}
\title{Update nodes to a g6 graph via proxy}
\usage{
g6_update_nodes(graph, nodes)
}
\arguments{
\item{graph}{A g6_proxy object created with \code{\link{g6_proxy}}.}

\item{nodes}{A data frame or list specifying the nodes to be updated.
If a data frame is provided, each row will be converted to a node.
Each node should have at least an 'id' field, and may include other attributes
such as 'label', 'style', 'x', 'y', etc. based on G6 node specifications.}
}
\value{
The g6_proxy object (invisibly), allowing for method chaining.
}
\description{
This function updates one or more nodes to an existing g6 graph instance
using a proxy object. This allows updating the graph without completely
re-rendering it.
}
\details{
This function can only be used with a g6_proxy object within a Shiny application.
It will not work with regular g6 objects outside of Shiny.

See \url{https://g6.antv.antgroup.com/en/api/data#graphupdatenodedata} for more details.
}
\seealso{
\code{\link{g6_proxy}}, \code{\link{g6_remove_nodes}}
}
