% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{g6R-package}
\alias{g6R}
\alias{g6R-package}
\title{g6R: Graph Visualisation Engine Widget for R and 'shiny' Apps}
\description{
Create stunning network experiences powered by the 'G6' graph visualisation engine 'JavaScript' library \url{https://g6.antv.antgroup.com/en}. In 'shiny' mode, modify your graph directly from the server function to dynamically interact with nodes and edges. Select your favorite layout among 20 choices. 15 behaviors are available such as interactive edge creation, collapse-expand and brush select. 17 plugins designed to improve the user experience such as a mini-map, toolbars and grid lines. Customise the look and feel of your graph with comprehensive options for nodes, edges and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/g6R}
  \item \url{https://cynkra.github.io/g6R/}
  \item Report bugs at \url{https://github.com/cynkra/g6R/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

Authors:
\itemize{
  \item David Schoch \email{david@schochastics.net}
}

Other contributors:
\itemize{
  \item  cynkra GmbH [funder]
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
