% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{is_g6_data}
\alias{is_g6_data}
\title{Check if an object is a g6_data object}
\usage{
is_g6_data(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
Logical indicating if \code{x} is of class \code{g6_data}.
}
\description{
Check if an object is a g6_data object
}
