% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{is_g6_node}
\alias{is_g6_node}
\alias{is_g6_edge}
\alias{is_g6_combo}
\title{Check if an object is a G6 element}
\usage{
is_g6_node(x)

is_g6_edge(x)

is_g6_combo(x)
}
\arguments{
\item{x}{An object of class \code{g6_node}, \code{g6_edge}, or \code{g6_combo}.}
}
\value{
Logical indicating if \code{x} is of the specified class.
}
\description{
Check if an object is a G6 element
}
