% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{is_g6_nodes}
\alias{is_g6_nodes}
\alias{is_g6_edges}
\alias{is_g6_combos}
\title{Check if an object is a list of G6 elements}
\usage{
is_g6_nodes(x)

is_g6_edges(x)

is_g6_combos(x)
}
\arguments{
\item{x}{An object of class \code{g6_nodes}, \code{g6_edges}, or \code{g6_combos}.}
}
\description{
Check if an object is a list of G6 elements
}
