% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{JS}
\alias{JS}
\title{Marks as string to be processed as a JS function}
\usage{
JS(...)
}
\arguments{
\item{...}{Any valid JS element.}
}
\value{
A character vector with class "JS_EVAL" that can be used in htmlwidgets
to mark is as a JavaScript function.
}
\description{
Useful for htmlwidgets
}
