% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{mds_layout}
\alias{mds_layout}
\title{Generate G6 MDS layout configuration}
\usage{
mds_layout(linkDistance = NULL, center = NULL, ...)
}
\arguments{
\item{linkDistance}{Ideal edge length (px).}

\item{center}{Centre of the layout, a vector of x and y.}

\item{...}{Additional parameters passed to the layout.
See \url{https://g6.antv.antgroup.com/en/manual/layout/mds-layout}.}
}
\value{
A list containing the layout configuration.
}
\description{
Multidimensional scaling: places nodes so that drawn distances follow
graph distances.
}
\examples{
mds_layout(linkDistance = 80)
}
