% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poke}
\alias{poke}
\title{Example pokemon data}
\format{
A list with 3 nested lists:
\describe{
\item{nodes}{list with 120 pokemon.}
\item{edges}{list with 69 connections.}
\item{combo}{list with 51 pokemon families.
A family contains all evolutions of a pokemon.}
}
}
\source{
\url{https://pokeapi.co/docs/v2}
}
\usage{
data(poke)
}
\description{
Pokemon evolution network to showcase combo features.
}
\keyword{datasets}
