% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radial}
\alias{radial}
\title{Example Network for radial layouts}
\format{
A list with 2 data frames:
\describe{
\item{nodes}{data frame with 34 rows for the nodes.}
\item{edges}{data frame with 58 rows for the edges.}
}
}
\source{
\url{https://assets.antv.antgroup.com/g6/radial.json}
}
\usage{
data(radial)
}
\description{
Example Network for radial layouts
}
\keyword{datasets}
