% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{random_layout}
\alias{random_layout}
\title{Generate G6 Random layout configuration}
\usage{
random_layout(width = NULL, height = NULL, ...)
}
\arguments{
\item{width, height}{Size of the area nodes are placed in (px).}

\item{...}{Additional parameters passed to the layout.
See \url{https://g6.antv.antgroup.com/en/manual/layout/random-layout}.}
}
\value{
A list containing the layout configuration.
}
\description{
Generate G6 Random layout configuration
}
\examples{
random_layout()
}
