% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{shinyG6-package}
\alias{shinyG6}
\alias{shinyG6-package}
\title{shinyG6: Graph Visualisation Engine Widget for R and 'shiny' Apps}
\description{
Create stunning network experiences powered by the G6 graph visualisation engine JavaScript library \url{https://g6.antv.antgroup.com/en}. In 'shiny' mode, modify your graph directly from the server function to dynamically interact nodes and edges. Select you favorite layout among 20 choices. 15 behaviors are available such as interactive edge creation, collapse-expand and brush select. 17 plugins designed to improve the user experience such as a minimap, toolbars and grid lines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/shinyG6}
  \item Report bugs at \url{https://github.com/cynkra/shinyG6/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

}
\keyword{internal}
