% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Example tree graph}
\format{
A list with 2 data frames:
\describe{
\item{nodes}{data frame with 31 rows for the nodes.}
\item{edges}{data frame with 30 rows for the edges.}
}
}
\source{
\url{https://gw.alipayobjects.com/os/antvdemo/assets/data/algorithm-category.json}
}
\usage{
data(tree)
}
\description{
The graph contains a classification of algorithm categories
}
\keyword{datasets}
