% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{validate_edges_ports}
\alias{validate_edges_ports}
\title{Validate that all edges connect input/output ports}
\usage{
validate_edges_ports(edges, nodes)
}
\arguments{
\item{edges}{List of g6_edge objects.}

\item{nodes}{List of g6_node objects.}
}
\value{
Invisibly TRUE if all edges are valid, otherwise stops with an error.
}
\description{
Validate that all edges connect input/output ports
}
