% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{validate_port}
\alias{validate_port}
\alias{validate_port.g6_port}
\title{Validate a single G6 port}
\usage{
validate_port(x, ...)

\method{validate_port}{g6_port}(x, ...)
}
\arguments{
\item{x}{An object of class 'g6_port'.}
}
\value{
The validated port (invisibly).
}
\description{
Validate a single G6 port
}
\examples{
validate_port(g6_port("input-1", type = "input"))
}
