% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{validate_ports}
\alias{validate_ports}
\alias{validate_ports.list}
\alias{validate_ports.g6_ports}
\title{Validate a list of G6 ports}
\usage{
validate_ports(x, ...)

\method{validate_ports}{list}(x, ...)

\method{validate_ports}{g6_ports}(x, ...)
}
\arguments{
\item{x}{A list of g6_port objects.}
}
\value{
The validated list (invisibly).
}
\description{
Validate a list of G6 ports
}
\examples{
validate_ports(list(
  g6_port("input-1", type = "input"),
  g6_port("output-1", type = "output")
))
}
