% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{g6_data}
\alias{g6_data}
\alias{as_g6_data}
\alias{as_g6_data.g6_data}
\alias{as_g6_data.list}
\title{Create a g6_data object}
\usage{
g6_data(nodes = NULL, edges = NULL, combos = NULL)

as_g6_data(x, ...)

\method{as_g6_data}{g6_data}(x, ...)

\method{as_g6_data}{list}(x, ...)
}
\arguments{
\item{nodes}{Nodes data.}

\item{edges}{Edges data.}

\item{combos}{Combo data.}

\item{x}{A list with elements \code{nodes}, \code{edges}, and/or \code{combos}.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{g6_data}.
}
\description{
Create compatible data structure for \code{\link[=g6_add_data]{g6_add_data()}}, \code{\link[=g6_set_data]{g6_set_data()}} or
simply \code{\link[=g6]{g6()}}.
}
